fix signedness warning

+++ Source/MemoryBuffer.hpp
@@ -15,7 +15,7 @@ public:
     memset(buffer, 0, size*channels*sizeof(float));
   }
   FloatArray getSamples(int channel) override {
-    ASSERT(channel < channels, "Invalid channel");
+    ASSERT((size_t)channel < channels, "Invalid channel");
     return FloatArray(buffer+channel*size, size);
   }
   int getChannels() override {
